remove code duplication of bb8 pool
